Move items to shared. Most of these will be used either by the analyzer or the search service.
