+++ app/lib/shared/model_properties.dart
@@ -1,0 +1,59 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:gcloud/db.dart';
+
+/// Similar to [ListProperty] but one which is fully compatible with python's
+/// 'db' implementation.
+class CompatibleListProperty extends Property {
+  final PrimitiveProperty subProperty;
+
+  const CompatibleListProperty(this.subProperty,
+      {String propertyName, bool indexed: true})
+      : super(propertyName: propertyName, required: true, indexed: indexed);
+
+  @override
+  bool validate(ModelDB db, Object value) {
+    if (!super.validate(db, value) || value is! List) return false;
+
+    for (var entry in value) {
+      if (!subProperty.validate(db, entry)) return false;
+    }
+    return true;
+  }
+
+  @override
+  Object encodeValue(ModelDB db, Object value, {bool forComparison: false}) {
+    if (forComparison) {
+      return subProperty.encodeValue(db, value, forComparison: true);
+    }
+
+    // NOTE: As opposed to [ListProperty] we will encode
+    //    - `null` as `[]`  (as opposed to `null`)
+    //    - `[]` as `[]`  (as opposed to `null`)
+    //    - `[a]` as `[a]` (as opposed to `a`)
+    if (value == null) return [];
+    final List list = value;
+    if (list.length == 0) return [];
+    if (list.length == 1) return [subProperty.encodeValue(db, list[0])];
+    return list.map((value) => subProperty.encodeValue(db, value)).toList();
+  }
+
+  @override
+  Object decodePrimitiveValue(ModelDB db, Object value) {
+    if (value == null) return [];
+    if (value is! List) return [subProperty.decodePrimitiveValue(db, value)];
+    return (value as List)
+        .map((entry) => subProperty.decodePrimitiveValue(db, entry))
+        .toList();
+  }
+}
+
+/// Similar to [StringListProperty] but one which is fully compatible with
+/// python's 'db' implementation.
+class CompatibleStringListProperty extends CompatibleListProperty {
+  const CompatibleStringListProperty({String propertyName, bool indexed: true})
+      : super(const StringProperty(required: true),
+            propertyName: propertyName, indexed: indexed);
+}
